provide JSQLBaseVisitor
